Move workspaceagent endpoint
